requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
SQLAlchemy==1.3.10
flask_cors==3.0.8
ConfigParser==5.0.0
marshmallow-sqlalchemy==0.23.1
passlib==1.7.2
flask-httpauth==4.1.0
